+++ DataswornPoco/csharp-system-text/AssetConditionMeter.cs
@@ -14,10 +14,10 @@ namespace Dataforged
         public string Label { get; set; }
 
         [JsonPropertyName("max")]
-        public sbyte Max { get; set; }
+        public int Max { get; set; }
 
         [JsonPropertyName("min")]
-        public sbyte Min { get; set; }
+        public int Min { get; set; }
 
         [JsonPropertyName("controls")]
         [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
@@ -25,6 +25,6 @@ namespace Dataforged
 
         [JsonPropertyName("value")]
         [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
-        public sbyte? Value { get; set; }
+        public int? Value { get; set; }
     }
 }
