+++ DataswornPoco/csharp-system-text/AssetOptionFieldChoicesNumberChoice.cs
@@ -13,7 +13,7 @@ namespace Dataforged
         public string Label { get; set; }
 
         [JsonPropertyName("value")]
-        public sbyte Value { get; set; }
+        public int Value { get; set; }
 
         [JsonPropertyName("selected")]
         [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
